new file

+++ Desktop/RampModel.cpp
@@ -163,7 +163,7 @@ bool Ramp::Model::setData(const QModelIndex& index, const QVariant& value, int r
    }
    else if (Core::Target::PolyRampLength == target)
    {
-      const uint8_t length = value.toInt();
+      const uint32_t length = value.toInt();
       polyRamp->setLength(length);
 
       callOnAllInstances(&Interface::modelHasChanged, identifier);
